+++ src/time_control.c
@@ -6,7 +6,7 @@
 /*   By: devrafaelly <devrafaelly@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/02 20:37:43 by devrafaelly       #+#    #+#             */
-/*   Updated: 2026/02/07 14:09:36 by devrafaelly      ###   ########.fr       */
+/*   Updated: 2026/02/07 15:50:38 by devrafaelly      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 #include <sys/time.h>
 
-long	get_current_time(void)
+long long	get_current_time(void)
 {
 	struct timeval	tv;
 
@@ -23,7 +23,7 @@ long	get_current_time(void)
 	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
 
-long	timestamp(t_data *data)
+long long	timestamp(t_data *data)
 {
 	return (get_current_time() - data->start_time);
 }
